Update physics checks
- Sphere-cast
- No more hardcoded gridsizes

+++ Packages/LMCore/Runtime/Crawler/GridEntityController.cs
@@ -4,6 +4,6 @@ namespace LMCore.Crawler
 {
     public interface GridEntityController
     {
-        public bool CanMoveTo(Movement movement, int length);
+        public bool CanMoveTo(Movement movement, float length);
     }
 }
